requirements.txt
certifi==2024.8.30
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
kiwisolver==1.4.5
lz4==4.0.2
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pillow==11.0.0
pyparsing==3.1.2
pypng==0.0.20
python-dateutil==2.9.0.post0
setuptools==75.2.0
six==1.16.0
stringcase==1.2.0
wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.7_1/libexec/wheel-0.44.0-py3-none-any.whl#sha256=c01326144883c50263d6967d1145640f3bbc8dfab32c2cbe817ef61d7a810219
